Develop

+++ api_blueprints/show_episodes.py
@@ -5,12 +5,13 @@ from sqlalchemy.orm import Session
 
 from database import engine
 from database.models import ShowEpisode
+from database.models import User
 
 show_episodes_blueprint = Blueprint("show_episodes_blueprint", __name__)
 
 CORS(show_episodes_blueprint, supports_credentials=True)
 
-@show_episodes_blueprint.route("/<int:id>", methods=['PUT'])
+@show_episodes_blueprint.route("/<int:id>", methods=["PUT"])
 @jwt_required()
 def update_show_episode(id: int):
     session = Session(engine)
@@ -21,16 +22,21 @@ def update_show_episode(id: int):
         session.close()
         return updated_episode_dict
     session.close()
-    return { 'message': f"This episode could not be found" }, 404
+    return { "message": f"This episode could not be found" }, 404
 
-@show_episodes_blueprint.route("/<int:id>", methods=['DELETE'])
+@show_episodes_blueprint.route("/<int:id>", methods=["DELETE"])
 @jwt_required()
 def delete_show_episode(id: int):
     session = Session(engine)
+    
+    current_user: User = get_current_user()
+    if current_user.role != "Admin":
+        return { "message": "You are not authorised to delete episodes" }, 403
+
     episode = ShowEpisode.get_episode_by_id(id, session)
     if episode:
         episode.delete_episode(session)
         session.close()
-        return '', 204
+        return "", 204
     session.close()
-    return { 'message': f"This episode could not be found" }, 404
+    return { "message": f"This episode could not be found" }, 404
